refactor/#223 : 알림 로직을 리팩토링한다
