+++ frontend/src/fcmConfig.ts
@@ -1,5 +1,4 @@
 import { initializeApp } from 'firebase/app';
-import { getAnalytics } from 'firebase/analytics';
 import { getMessaging } from 'firebase/messaging';
 
 const firebaseConfig = {
@@ -13,5 +12,4 @@ const firebaseConfig = {
 };
 
 const app = initializeApp(firebaseConfig);
-const analytics = getAnalytics(app);
 export const messaging = getMessaging(app);
