+++ frontend/src/pages/workspace/WorkspaceListPage.tsx
@@ -4,49 +4,14 @@ import useUserWorkspacesQuery from '@/hooks/workspace/useUserWorkspacesQuery';
 import WorkspaceListItem from '@/components/workspace/WorkspaceListItem';
 import { useNavigate } from 'react-router';
 import { userOriginStore } from '@/stores/userOriginStore';
-import { useEffect } from 'react';
-import { messaging } from '@/firebase-messaging-sw';
-import { getToken } from 'firebase/messaging';
-import { postFirebaseToken } from '@/apis/alarm';
+import { usePushTokenInit } from '@/hooks/alarm.ts/usePushTokenInit';
 
 const WorkSpaceListPage = () => {
   const navigate = useNavigate();
   const { workspacesInfo, isWorkspacesError, isWorkspacesLoading } =
     useUserWorkspacesQuery();
   const { user } = userOriginStore();
-
-  useEffect(() => {
-    if (!user?.id) return;
-    Notification.requestPermission()
-      .then(permission => {
-        if (permission === 'granted') {
-          getToken(messaging, {
-            vapidKey: import.meta.env.VITE_FIREBASE_VAPID_KEY,
-          }).then(fmc_responsed_token => {
-            const expiredAt = new Date();
-            expiredAt.setDate(expiredAt.getDate() + 30); // 30일 후 만료일 설정
-            const exsiting_token = localStorage.getItem('fcm_token');
-            // localstorage에 없거나 있어도 fcm에서 발급받은것과 저장된것이 같다면 서버에 업데이트 해줄 필요가 없음
-            console.log(fmc_responsed_token);
-
-            if (!exsiting_token || fmc_responsed_token !== exsiting_token) {
-              return postFirebaseToken({
-                userId: user.id,
-                token: fmc_responsed_token,
-                expiredAt: expiredAt.toISOString(),
-              }).then(() => {
-                localStorage.setItem('fcm_token', fmc_responsed_token);
-              });
-            }
-          });
-        } else {
-          alert('알림 설정을 거부하셨습니다.');
-        }
-      })
-      .catch(err => {
-        console.error('에러 log', err);
-      });
-  }, []);
+  usePushTokenInit({ user });
 
   if (isWorkspacesLoading) return <div>로딩중...</div>;
   if (isWorkspacesError) return <div>에러 발생</div>;
